requirements.txt
pandas==0.24.2
xlrd==1.2.0
jinja2==2.10
python-dotenv==0.14.0
